debugging

+++ models/cocsp.py
@@ -63,7 +63,7 @@ class CoCSPInterface(CLIPInterface):
     ):
         soft_embeddings = soft_embeddings.unsqueeze(0).repeat(config.train_batch_size, 1, 1)
         super().__init__(
-            clip_model.half(),
+            clip_model,
             config,
             class_token_ids,
             soft_embeddings,
@@ -75,6 +75,8 @@ class CoCSPInterface(CLIPInterface):
         self.offset = offset
         self.attr_dropout = nn.Dropout(attr_dropout)
         self.vctx_encoder = vctx_encoder.to(self.device)
+        self.text_encoder = self.text_encoder.half()
+        self.clip_model.visual = self.clip_model.visual.half()
 
     def construct_token_tensors(self, batch_img, pair_idx):
